requirements.txt
transformers>=4.41.0
trl>=0.9.0
peft>=0.11.0
accelerate>=0.30.0
bitsandbytes>=0.43.0
datasets>=2.19.0
evaluate>=0.4.2
scikit-learn>=1.4.2
sentencepiece>=0.2.0
pyyaml>=6.0.1
fastapi>=0.111.0
uvicorn>=0.30.0
wandb>=0.17.0
